Inline trace errors.
We currently have two trace error enums, only one of which is in
`errors.rs`. Inlining both of them in `mod.rs` feels more idiomatic.
