+++ ykrt/src/trace/swt/mod.rs
@@ -1,6 +1,6 @@
 //! Software tracer.
 
-use super::{errors::InvalidTraceError, AOTTraceIterator, TraceAction, TraceRecorder};
+use super::{AOTTraceIterator, InvalidTraceError, TraceAction, TraceRecorder};
 use crate::{compile::jitc_llvm::frame::BitcodeSection, mt::DEFAULT_TRACE_TOO_LONG};
 use std::sync::Once;
 use std::{cell::RefCell, collections::HashMap, error::Error, ffi::CString, sync::Arc};
